Add files via upload
"""Pree says he stole Daenerys' dragons, and would return them to Daenerys only if she can pair the dragons in a peculiar way.
There are N (N is even) dragons, the strength of ith dragon is Ai. She needs to create N/2 pairs of the dragons. The strength of the dragon pair is the sum of the strengths of dragons in the pair. 
She needs to pair them in such a way that the difference between the strength of dragon pair with maximum strength and the strength of the dragon pair with minimum strength is minimised.

Now, you need to find the minimum difference in strengths if she has paired them correctly.
Input
The first line of the input contains an integer N, the number of dragons.
The second line of the input contains N integers, the strengths of the dragons.


Constraints
1 <= N <= 200000 (so many dragons, huh)
1 <= Ai <= 109 for all values of i
N is divisible by 2 """